Move grpc query and msg server to pricefeed root
